# perf: use faster uuid if available