--- a/packages/common/package.json
+++ b/packages/common/package.json
@@ -22,6 +22,9 @@
     "tslib": "2.3.1",
     "uuid": "8.3.2"
   },
+  "optionalDependencies": {
+    "@napi-rs/uuid": "^0.2.0"
+  },
   "peerDependencies": {
     "cache-manager": "*",
     "class-transformer": "*",