--- a/packages/common/utils/random-string-generator.util.ts
+++ b/packages/common/utils/random-string-generator.util.ts
@@ -1,3 +1,13 @@
-import { v4 as uuid } from 'uuid';
+import { v4 } from 'uuid';
+
+const uuid = (function () {
+  try {
+    // eslint-disable-next-line @typescript-eslint/no-var-requires
+    return require('@napi-rs/uuid').v4;
+  } catch (e) {
+    // for non-supported platforms, fallback to JavaScript implementation
+    return v4;
+  }
+})();
 
 export const randomStringGenerator = () => uuid();