--- a/packages/core/injector/modules-container.ts
+++ b/packages/core/injector/modules-container.ts
@@ -1,5 +1,15 @@
-import { v4 as uuid } from 'uuid';
+import { v4 } from 'uuid';
 import { Module } from './module';
+
+const uuid = (function () {
+  try {
+    // eslint-disable-next-line @typescript-eslint/no-var-requires
+    return require('@napi-rs/uuid').v4;
+  } catch (e) {
+    // for non-supported platforms, fallback to JavaScript implementation
+    return v4;
+  }
+})();
 
 export class ModulesContainer extends Map<string, Module> {
   private readonly _applicationId = uuid();