--- a/packages/core/middleware/utils.ts
+++ b/packages/core/middleware/utils.ts
@@ -3,9 +3,19 @@ import { HttpServer, RouteInfo, Type } from '@nestjs/common/interfaces';
 import { isFunction } from '@nestjs/common/utils/shared.utils';
 import { iterate } from 'iterare';
 import * as pathToRegexp from 'path-to-regexp';
-import { v4 as uuid } from 'uuid';
+import { v4 } from 'uuid';
 import { ExcludeRouteMetadata } from '../router/interfaces/exclude-route-metadata.interface';
 import { isRouteExcluded } from '../router/utils';
+
+const uuid = (function () {
+  try {
+    // eslint-disable-next-line @typescript-eslint/no-var-requires
+    return require('@napi-rs/uuid').v4;
+  } catch (e) {
+    // for non-supported platforms, fallback to JavaScript implementation
+    return v4;
+  }
+})();
 
 export const mapToExcludeRoute = (
   routes: RouteInfo[],