--- a/packages/core/package.json
+++ b/packages/core/package.json
@@ -35,6 +35,9 @@
     "tslib": "2.3.1",
     "uuid": "8.3.2"
   },
+  "optionalDependencies": {
+    "@napi-rs/uuid": "^0.2.0"
+  },
   "devDependencies": {
     "@nestjs/common": "8.2.5"
   },