## PR Checklist
Please check if your PR fulfills the following requirements:

- [x] The commit message follows our guidelines: https://github.com/nestjs/nest/blob/master/CONTRIBUTING.md
- [ ] Tests for the changes have been added (for bug fixes / features)
- [x] Docs have been added / updated (for bug fixes / features)


## PR Type
What kind of change does this PR introduce?

<!-- Please check the one that applies to this PR using "x". -->
- [ ] Bugfix
- [ ] Feature
- [ ] Code style update (formatting, local variables)
- [ ] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] CI related changes
- [x] Other... Please describe:

## Other information

https://github.com/Brooooooklyn/uuid#performance

```
Running "UID string" suite...
Progress: 100%

  crypto.randomUUID:
    810 767 ops/s, ±0.50%      | slowest, 92.69% slower

  napi-uuid-v4:
    11 084 015 ops/s, ±0.43%   | fastest

  @lukeed/uuid:
    7 506 916 ops/s, ±0.69%    | 32.27% slower

  nanoid:
    5 736 698 ops/s, ±0.54%    | 48.24% slower

  uuid-js:
    1 855 307 ops/s, ±0.30%    | 83.26% slower

Finished 5 cases!
  Fastest: napi-uuid-v4
  Slowest: crypto.randomUUID
```