@kamilmysliwiec 2022-01-19
Thanks for your contribution @Brooooooklyn.

I appreciate your work on the uuid package but the performance difference is so minor (given how we even use this package) that it literally makes zero sense to migrate.